also move LegendHelper to Models
