+++ lib/Models/LegendHelper.js
