+++ lib/Models/RegionMapping.js
@@ -16,7 +16,7 @@ var when = require('terriajs-cesium/Source/ThirdParty/when');
 var arraysAreEqual = require('../Core/arraysAreEqual');
 var getUniqueValues = require('../Core/getUniqueValues');
 var ImageryProviderHooks = require('../Map/ImageryProviderHooks');
-var LegendHelper = require('../Map/LegendHelper');
+var LegendHelper = require('../Models/LegendHelper');
 var RegionProviderList = require('../Map/RegionProviderList');
 var TableStructure = require('../Map/TableStructure');
 var TableStyle = require('../Models/TableStyle');
