+++ lib/Models/TableDataSource.js
@@ -20,7 +20,7 @@ var TimeInterval = require('terriajs-cesium/Source/Core/TimeInterval');
 var TimeIntervalCollectionProperty = require('terriajs-cesium/Source/DataSources/TimeIntervalCollectionProperty');
 var VerticalOrigin = require('terriajs-cesium/Source/Scene/VerticalOrigin');
 
-var LegendHelper = require('../Map/LegendHelper');
+var LegendHelper = require('../Models/LegendHelper');
 var TableStructure = require('../Map/TableStructure');
 var TableStyle = require('../Models/TableStyle');
 var VarType = require('../Map/VarType');
